Added files via upload
This script was written to be incorporated within an existing mapping system. 
Python was chosen as a simple way to get the project off the ground. 
The script was called by a PHP script from an existing website. 
The script can be run independently when provided with arguments. 
This script will not run without the appropriate databases that were used when it was developed.